+++ src/afterfact.rs
@@ -502,15 +502,23 @@ fn calc_statistic_info(
         afterfact_info
             .timestamps
             .push(detect_info.detected_time.timestamp());
-        if !detect_info.is_condition {
-            afterfact_info
-                .detected_record_idset
-                .insert(CompactString::from(format!(
-                    "{}_{}",
-                    detect_info.detected_time, detect_info.eventid
-                )));
+        match &detect_info.agg_result {
+            None => {
+                afterfact_info
+                    .detected_record_idset
+                    .insert(CompactString::from(format!(
+                        "{}_{}",
+                        detect_info.detected_time, detect_info.eventid
+                    )));
+            }
+            Some(agg_result) => {
+                agg_result.agg_record_time_info.iter().for_each(|a| {
+                    afterfact_info
+                        .detected_record_idset
+                        .insert(CompactString::from(format!("{}_{}", a.time, a.event_id)));
+                });
+            }
         }
-
         if !output_option.no_summary {
             let level_suffix = get_level_suffix(detect_info.level.as_str());
             let author_list = afterfact_info
@@ -546,23 +554,33 @@ fn calc_statistic_info(
                     .insert(detect_info.ruleid.to_owned());
                 afterfact_info.unique_detect_counts_by_level[level_suffix] += 1;
             }
-
-            let computer_rule_check_key = CompactString::from(format!(
-                "{}|{}",
-                &detect_info.computername, &detect_info.rulepath
-            ));
-            if !afterfact_info
-                .detected_computer_and_rule_names
-                .contains(&computer_rule_check_key)
-            {
-                afterfact_info
+            let computer_names = match &detect_info.agg_result {
+                None => vec![detect_info.computername.clone()],
+                Some(agg) => agg
+                    .agg_record_time_info
+                    .iter()
+                    .map(|a| CompactString::from(a.computer.clone()))
+                    .collect::<std::collections::HashSet<_>>() // Convert to HashSet to remove duplicates
+                    .into_iter()
+                    .sorted()
+                    .collect(),
+            };
+            for computername in &computer_names {
+                let computer_rule_check_key =
+                    CompactString::from(format!("{}|{}", computername, &detect_info.rulepath));
+                if !afterfact_info
                     .detected_computer_and_rule_names
-                    .insert(computer_rule_check_key);
-                countup_aggregation(
-                    &mut afterfact_info.detect_counts_by_computer_and_level,
-                    &detect_info.level,
-                    &detect_info.computername,
-                );
+                    .contains(&computer_rule_check_key)
+                {
+                    afterfact_info
+                        .detected_computer_and_rule_names
+                        .insert(computer_rule_check_key);
+                    countup_aggregation(
+                        &mut afterfact_info.detect_counts_by_computer_and_level,
+                        &detect_info.level,
+                        computername,
+                    );
+                }
             }
             afterfact_info.rule_title_path_map.insert(
                 detect_info.ruletitle.to_owned(),
@@ -579,6 +597,7 @@ fn calc_statistic_info(
                 &detect_info.level,
                 &detect_info.ruletitle,
             );
+            let level_suffix = get_level_suffix(detect_info.level.as_str());
             afterfact_info.total_detect_counts_by_level[level_suffix] += 1;
         }
     }
@@ -1901,7 +1920,7 @@ pub fn output_json_str(
                     let mut children_output_stock: HashMap<CompactString, Vec<CompactString>> =
                         HashMap::new();
                     let mut children_output_order = vec![];
-                    if detect_info.is_condition {
+                    if detect_info.agg_result.is_some() {
                         if details_target_stock[0] == "-" {
                             output_stock.push(_create_json_output_format(
                                 key,
@@ -2477,7 +2496,7 @@ mod tests {
                     eventid: CompactString::from(test_eventid),
                     detail: CompactString::default(),
                     ext_field: output_profile.to_owned(),
-                    is_condition: false,
+                    agg_result: None,
                     details_convert_map: HashMap::default(),
                 },
                 &profile_converter,
@@ -2501,7 +2520,7 @@ mod tests {
                     eventid: CompactString::from(test_eventid),
                     detail: CompactString::default(),
                     ext_field: output_profile.to_owned(),
-                    is_condition: false,
+                    agg_result: None,
                     details_convert_map: HashMap::default(),
                 },
                 &profile_converter,
@@ -2824,7 +2843,7 @@ mod tests {
                     eventid: CompactString::from(test_eventid),
                     detail: CompactString::default(),
                     ext_field: output_profile.to_owned(),
-                    is_condition: false,
+                    agg_result: None,
                     details_convert_map: HashMap::default(),
                 },
                 &profile_converter,
@@ -2848,7 +2867,7 @@ mod tests {
                     eventid: CompactString::from(test_eventid),
                     detail: CompactString::default(),
                     ext_field: output_profile.to_owned(),
-                    is_condition: false,
+                    agg_result: None,
                     details_convert_map: HashMap::default(),
                 },
                 &profile_converter,
@@ -3151,7 +3170,7 @@ mod tests {
                     eventid: CompactString::from(test_eventid),
                     detail: CompactString::default(),
                     ext_field: output_profile.to_owned(),
-                    is_condition: false,
+                    agg_result: None,
                     details_convert_map: HashMap::default(),
                 },
                 &profile_converter,
@@ -3175,7 +3194,7 @@ mod tests {
                     eventid: CompactString::from(test_eventid),
                     detail: CompactString::default(),
                     ext_field: output_profile.to_owned(),
-                    is_condition: false,
+                    agg_result: None,
                     details_convert_map: HashMap::default(),
                 },
                 &profile_converter,
@@ -3488,7 +3507,7 @@ mod tests {
                     eventid: CompactString::from(test_eventid),
                     detail: CompactString::default(),
                     ext_field: output_profile.to_owned(),
-                    is_condition: false,
+                    agg_result: None,
                     details_convert_map,
                 },
                 &profile_converter,
@@ -3512,7 +3531,7 @@ mod tests {
                     eventid: CompactString::from(test_eventid),
                     detail: CompactString::default(),
                     ext_field: output_profile.to_owned(),
-                    is_condition: false,
+                    agg_result: None,
                     details_convert_map: HashMap::default(),
                 },
                 &profile_converter,
@@ -3897,7 +3916,7 @@ mod tests {
                     eventid: CompactString::from(test_eventid),
                     detail: CompactString::default(),
                     ext_field: output_profile.to_owned(),
-                    is_condition: false,
+                    agg_result: None,
                     details_convert_map,
                 },
                 &profile_converter,
@@ -4251,7 +4270,7 @@ mod tests {
                     eventid: CompactString::from(test_eventid),
                     detail: CompactString::default(),
                     ext_field: output_profile.to_owned(),
-                    is_condition: false,
+                    agg_result: None,
                     details_convert_map,
                 },
                 &profile_converter,
@@ -4531,7 +4550,7 @@ mod tests {
                     eventid: CompactString::from(test_eventid),
                     detail: CompactString::default(),
                     ext_field: output_profile.to_owned(),
-                    is_condition: false,
+                    agg_result: None,
                     details_convert_map,
                 },
                 &profile_converter,